requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
xgboost>=2.0.0
torch>=2.0.0
anthropic>=0.7.0
python-binance>=1.0.19
solana>=0.30.0
streamlit>=1.30.0
plotly>=5.18.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
asyncio>=3.4.3
pyyaml>=6.0.1
psycopg2-binary>=2.9.9
ta>=0.10.0
pytest>=7.4.0
jupyter>=1.0.0
python-telegram-bot>=20.0
ccxt>=4.1.0
web3>=6.0.0 